Backport encoder from S905 to S805

+++ libvpcodec.cpp
@@ -139,7 +139,7 @@ int vl_video_encoder_encode(vl_codec_handle_t codec_handle, vl_frame_type_t fram
         videoInput.coding_timestamp = handle->mNumInputFrames * 1000 / videoInput.frame_rate;  // in ms
         videoInput.fmt = AMVENC_NV21;
         videoInput.YCbCr[0] = (unsigned long)in;
-        videoInput.YCbCr[1] = (unsigned)(videoInput.YCbCr[0] + videoInput.height * videoInput.pitch);
+        videoInput.YCbCr[1] = (unsigned long)(videoInput.YCbCr[0] + videoInput.height * videoInput.pitch);
         videoInput.YCbCr[2] = 0;
         videoInput.canvas = 0xffffffff;
         videoInput.type = VMALLOC_BUFFER;
